+++ frontend/next.config.js
@@ -1,3 +1,3 @@
-const withSass = require("@zeit/next-sass");
+const withSass = require('@zeit/next-sass');
 
 module.exports = withSass();
