+++ web/404.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html><head><meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1" class="next-head"/><meta charSet="utf-8" class="next-head"/><title class="next-head">404: This page could not be found</title><link rel="preload" href="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/_app.js" as="script"/><link rel="preload" href="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" as="script"/><link rel="preload" href="/_next/static/chunks/commons.c50355a0525fd0e3c680.js" as="script"/><link rel="preload" href="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" as="script"/></head><body><div id="__next"><div style="color:#000;background:#fff;font-family:-apple-system, BlinkMacSystemFont, Roboto, &quot;Segoe UI&quot;, &quot;Fira Sans&quot;, Avenir, &quot;Helvetica Neue&quot;, &quot;Lucida Grande&quot;, sans-serif;height:100vh;text-align:center;display:flex;flex-direction:column;align-items:center;justify-content:center"><div><style>body { margin: 0 }</style><h1 style="display:inline-block;border-right:1px solid rgba(0, 0, 0,.3);margin:0;margin-right:20px;padding:10px 23px 10px 0;font-size:24px;font-weight:500;vertical-align:top">404</h1><div style="display:inline-block;text-align:left;line-height:49px;height:49px;vertical-align:middle"><h2 style="font-size:14px;font-weight:normal;line-height:inherit;margin:0;padding:0">This page could not be found<!-- -->.</h2></div></div></div></div><script id="__NEXT_DATA__" type="application/json">{"dataManager":"[]","props":{"pageProps":{"statusCode":404}},"page":"/_error","query":{},"buildId":"7hlGC6IyGrAtu7WrUFi9N","dynamicBuildId":false,"nextExport":true}</script><script async="" id="__NEXT_PAGE__/_app" src="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/_app.js"></script><script src="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" async=""></script><script src="/_next/static/chunks/commons.c50355a0525fd0e3c680.js" async=""></script><script src="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" async=""></script></body></html>
+<!DOCTYPE html><html><head><meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1" class="next-head"/><meta charSet="utf-8" class="next-head"/><title class="next-head">404: This page could not be found</title><link rel="preload" href="/_next/static/felAkQBd6VVD03oA2-SI4/pages/_app.js" as="script"/><link rel="preload" href="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" as="script"/><link rel="preload" href="/_next/static/chunks/commons.e304fba0b9256846e73d.js" as="script"/><link rel="preload" href="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" as="script"/></head><body><div id="__next"><div style="color:#000;background:#fff;font-family:-apple-system, BlinkMacSystemFont, Roboto, &quot;Segoe UI&quot;, &quot;Fira Sans&quot;, Avenir, &quot;Helvetica Neue&quot;, &quot;Lucida Grande&quot;, sans-serif;height:100vh;text-align:center;display:flex;flex-direction:column;align-items:center;justify-content:center"><div><style>body { margin: 0 }</style><h1 style="display:inline-block;border-right:1px solid rgba(0, 0, 0,.3);margin:0;margin-right:20px;padding:10px 23px 10px 0;font-size:24px;font-weight:500;vertical-align:top">404</h1><div style="display:inline-block;text-align:left;line-height:49px;height:49px;vertical-align:middle"><h2 style="font-size:14px;font-weight:normal;line-height:inherit;margin:0;padding:0">This page could not be found<!-- -->.</h2></div></div></div></div><script id="__NEXT_DATA__" type="application/json">{"dataManager":"[]","props":{"pageProps":{"statusCode":404}},"page":"/_error","query":{},"buildId":"felAkQBd6VVD03oA2-SI4","dynamicBuildId":false,"nextExport":true}</script><script async="" id="__NEXT_PAGE__/_app" src="/_next/static/felAkQBd6VVD03oA2-SI4/pages/_app.js"></script><script src="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" async=""></script><script src="/_next/static/chunks/commons.e304fba0b9256846e73d.js" async=""></script><script src="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" async=""></script></body></html>
