+++ web/index/index.html
@@ -1,1 +1,1 @@
-<!DOCTYPE html><html><head><meta charSet="utf-8" class="next-head"/><meta name="viewport" content="width=device-width,minimum-scale=1,initial-scale=1" class="next-head"/><link rel="preload" href="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/index.js" as="script"/><link rel="preload" href="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/_app.js" as="script"/><link rel="preload" href="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" as="script"/><link rel="preload" href="/_next/static/chunks/commons.c50355a0525fd0e3c680.js" as="script"/><link rel="preload" href="/_next/static/chunks/styles.37a297b15431f06f5fd7.js" as="script"/><link rel="preload" href="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" as="script"/><link rel="stylesheet" href="/_next/static/css/styles.e36ad40b.chunk.css"/></head><body><div id="__next"><div class="container"><h1 class="title">Last.fm Collage Generator</h1><p>Generate Last.fm collages from scrobble data!</p><form id="userForm"><label for="username">Last.fm Username</label><br/><input type="text" id="username" name="username" value=""/><select name="size" id="size"><option selected="" value="3">3x3</option><option value="4">4x4</option><option value="5">5x5</option><option value="6">6x6</option><option value="7">7x7</option></select><select name="time" id="time"><option selected="" value="7day">Week</option><option value="1month">Month</option><option value="3month">3 Month</option><option value="6month">6 Month</option><option value="12month">Year</option><option value="overall">Overall</option></select><button type="submit">SUBMIT</button></form><div class="bottom"><a href="https://github.com/NathanYocum/lastfm-collage-generator">View Source on GitHub!</a></div></div></div><script id="__NEXT_DATA__" type="application/json">{"dataManager":"[]","props":{"pageProps":{}},"page":"/","query":{},"buildId":"7hlGC6IyGrAtu7WrUFi9N","dynamicBuildId":false,"nextExport":true}</script><script async="" id="__NEXT_PAGE__/" src="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/index.js"></script><script async="" id="__NEXT_PAGE__/_app" src="/_next/static/7hlGC6IyGrAtu7WrUFi9N/pages/_app.js"></script><script src="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" async=""></script><script src="/_next/static/chunks/commons.c50355a0525fd0e3c680.js" async=""></script><script src="/_next/static/chunks/styles.37a297b15431f06f5fd7.js" async=""></script><script src="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" async=""></script></body></html>
+<!DOCTYPE html><html><head><meta charSet="utf-8" class="next-head"/><meta name="viewport" content="initial-scale=1.0, width=device-width" class="next-head"/><title class="next-head">Lastfm Collage Generator</title><link rel="preload" href="/_next/static/felAkQBd6VVD03oA2-SI4/pages/index.js" as="script"/><link rel="preload" href="/_next/static/felAkQBd6VVD03oA2-SI4/pages/_app.js" as="script"/><link rel="preload" href="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" as="script"/><link rel="preload" href="/_next/static/chunks/commons.e304fba0b9256846e73d.js" as="script"/><link rel="preload" href="/_next/static/chunks/styles.37a297b15431f06f5fd7.js" as="script"/><link rel="preload" href="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" as="script"/><link rel="stylesheet" href="/_next/static/css/styles.7295b2b5.chunk.css"/></head><body><div id="__next"><div class="container"><h1 class="title">Last.fm Collage Generator</h1><p>Generate Last.fm collages from scrobble data!</p><form class="userForm" id="userForm"><label id="usernameLabel" for="username">Last.fm Username<br/><input type="text" id="username" name="username" value=""/></label><select name="size" id="size"><option selected="" value="3">3x3</option><option value="4">4x4</option><option value="5">5x5</option><option value="6">6x6</option><option value="7">7x7</option></select><select name="time" id="time"><option selected="" value="7day">Week</option><option value="1month">Month</option><option value="3month">3 Month</option><option value="6month">6 Month</option><option value="12month">Year</option><option value="overall">Overall</option></select><button type="submit">SUBMIT</button></form><div class="bottom"><a href="https://github.com/NathanYocum/lastfm-collage-generator">View Source on GitHub!</a></div></div></div><script id="__NEXT_DATA__" type="application/json">{"dataManager":"[]","props":{"pageProps":{}},"page":"/","query":{},"buildId":"felAkQBd6VVD03oA2-SI4","dynamicBuildId":false,"nextExport":true}</script><script async="" id="__NEXT_PAGE__/" src="/_next/static/felAkQBd6VVD03oA2-SI4/pages/index.js"></script><script async="" id="__NEXT_PAGE__/_app" src="/_next/static/felAkQBd6VVD03oA2-SI4/pages/_app.js"></script><script src="/_next/static/runtime/webpack-8ed9452df514b4d17d80.js" async=""></script><script src="/_next/static/chunks/commons.e304fba0b9256846e73d.js" async=""></script><script src="/_next/static/chunks/styles.37a297b15431f06f5fd7.js" async=""></script><script src="/_next/static/runtime/main-fd73dbd6f398b7b886ec.js" async=""></script></body></html>
